Moved error functions to new module
